Move asar code to a common location
