+++ win32.js
@@ -5,9 +5,9 @@ var child = require('child_process')
 var mkdirp = require('mkdirp')
 var rimraf = require('rimraf')
 var ncp = require('ncp').ncp
-var asar = require('asar')
 var mv = require('mv')
 var rcedit = require('rcedit')
+var common = require('./common')
 
 module.exports = {
   createApp: function createApp (opts, electronApp, cb) {
@@ -87,7 +87,8 @@ function buildWinApp (opts, cb, newApp) {
       copy(newApp, finalPath, function moved (err) {
         if (err) return cb(err)
         if (opts.asar) {
-          asarApp(function (err) {
+          var finalPath = path.join(opts.out || process.cwd(), opts.name + '-win32', 'resources')
+          common.asarApp(finalPath, function (err) {
             if (err) return cb(err)
             updateIcon()
           })
@@ -108,17 +109,6 @@ function buildWinApp (opts, cb, newApp) {
 
       rcedit(exePath, {icon: opts.icon}, function (err) {
         cb(err, finalPath)
-      })
-
-    }
-
-    function asarApp (cb) {
-      var finalPath = path.join(opts.out || process.cwd(), opts.name + '-win32', 'resources')
-      var src = path.join(finalPath, 'app')
-      var dest = path.join(finalPath, 'app.asar')
-      asar.createPackage(src, dest, function (err) {
-        if (err) return cb(err)
-        rimraf(src, cb)
       })
     }
   })
